Indian Cuisine Analysis
In India there are variety of dishes based on the different regions in each state. Food makers use local variety spices, herbs, dairy, vegetables and fruits to make n number of dishes and bring authentic taste.
With my analysis I have created a dashboard in which you will find interesting facts and ingredients about Indian food and diet.

Data found in Kaggle.
Page 1 Visualizations:

1. Proportion of Veg Vs Non-Veg
Observation(fact): India is ranked top in the world with 38% of the total population being vegetarians

2. No of dishes based on meal course 

3. Proportion of cooking time and preparation time based on diet
Observation: Most of the dishes both vegetarian and non-vegetarian have preparation time and cooking time between 0 to 120 minutes.

4. No of dishes based on region by diet
Observation: About 129 dies are eaten as the main course. Chicken Tikka Masala which is most favorite dish for many Indians is from the roots of Scotland.   

5. Percentage of dishes by flavor profile
Observation: More than 58% of the dishes are spicy in flavor.
India is the largest producer of spices in the world. According to Ayurveda, Indian food is a balance of 6 flavors: sweet, sour, salty, bitter, spicy, and astringent. Each flavor is imparted from a particular ingredient in a dish. This is the uniqueness of Indian cuisine!

6. Word cloud shows the commonly used ingredients in Indian cuisine.

Page 2 Visualizations:

6. Top 10 snack items with longest cooking time

7. Top 10 main course dishes with longest cooking time based on state

8. South Indian food mini info-graph that shows percentage of vg and non veg population and total dishes 
per course in South India